requirements.txt
fastapi==0.115.12
firecrawl-py==2.2.0
llama-index==0.12.32
llama-index-llms-openai-like==0.3.4
lxml==5.4.0
playwright==1.51.0
pytest==8.3.5
python-dotenv==1.0.0
requests==2.32.3
retry==0.9.2
search_engines==1.0.7
uvicorn==0.34.2